Hexagonal

+++ microservices/measurement-based-care/mbc/domain/model/patient.py
@@ -2,11 +2,12 @@ from pydantic import BaseModel
 
 from mbc.domain.model.measurement import Measurement
 
+from decimal import Decimal
 
 class Patient(BaseModel):
     registry_id: str
     user_id: str
-    created_at: float
+    created_at: Decimal
     is_enabled: bool = True
 
     flags: dict[str, bool] = {}
@@ -15,7 +16,7 @@ class Patient(BaseModel):
 
     baseline_measurements: dict[str, Measurement] = {}
     latest_measurements: dict[str, Measurement] = {}
-    dates_of_interaction: dict[str, float] = {}
+    dates_of_interaction: dict[str, Decimal] = {}
 
     def billable(self, type: str, minutes: int):
         print("save billable event")
